Add runtime UID cache
When uploading a new entry the path already defines an UID but The
adapters don't have a way to store it and assign a new one. Different
clients handle this differently:

DAVx⁵ adopts to the new UID by deleting the old entry and creating a new
one. Thunderbird shows an error and Evolution keeps the old UID.

With the cache Thunderbird works as expected, even when restarting
Radicale. Evolution works but creates a new entry if Radicale is
restarted.

A solution could be to store the cache on disk or to store the UID with
the entry in the adapter.

Closes: #2